requirements.txt
langchain
langchain-community
streamlit
openai
faiss-cpu
python-dotenv
